Scale map on window resize

+++ include/map.h
@@ -1,6 +1,6 @@
 #ifndef MAP_
 #define MAP_
 
-void drawMap(const int startX, const int startY, const int sizeMod);
+void drawMap(const int startX, const int startY, const float sizeMod);
 
 #endif // MAP_
